tuna/utils/metadata.py created for general constant definitions get_miopen_v->worker.py is moved to fin_class->miopen/worker __init__.py added in miopen/worker directory
